requirements.txt
pyupbit==0.2.33
requests==2.31.0
websockets==12.0
pandas==2.2.0
numpy==1.26.4
pandas-ta==0.3.14b0
python-telegram-bot==20.7
apscheduler==3.10.4
pyyaml==6.0.1
PyJWT==2.8.0
cryptography==42.0.2
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.26.0
